requirements.txt
# Core Framework
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
scalar-fastapi>=1.8.0

# Data Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.21.0

# Database
SQLAlchemy>=2.0.0
asyncpg>=0.27.0
aiosqlite>=0.17.0
alembic>=1.11.0

# Security & Authentication
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0

# Optional: Caching & Rate Limiting
redis>=4.5.0

# Optional: Cloud Storage
boto3>=1.26.0  # AWS S3
google-cloud-storage>=2.10.0  # Google Cloud Storage

# Optional: Email
sendgrid>=6.10.0

# Utilities
python-multipart>=0.0.6
